validation added

+++ src/main/java/com/ba/restaurant/controller/CategoryController.java
@@ -2,10 +2,13 @@ package com.ba.restaurant.controller;
 
 import com.ba.restaurant.dto.CategoryDTO;
 import com.ba.restaurant.dto.ProductDTO;
+import com.ba.restaurant.exception.BusinessRuleException;
 import com.ba.restaurant.service.CategoryService;
+import com.ba.restaurant.exception.BusinessMessages;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @CrossOrigin(origins = "*")
@@ -17,7 +20,7 @@ public class CategoryController {
     CategoryService categoryService;
 
     @PostMapping("/add")
-    public CategoryDTO addCategory(@RequestBody CategoryDTO categoryDTO) {
+    public CategoryDTO addCategory(@Valid  @RequestBody CategoryDTO categoryDTO) {
         categoryService.addCategory(categoryDTO);
         return categoryDTO;
     }
@@ -28,23 +31,35 @@ public class CategoryController {
     }
 
     @PutMapping("/update/")
-    public CategoryDTO updateCategory(@RequestBody CategoryDTO categoryDTO) {
+    public CategoryDTO updateCategory(@Valid @RequestBody CategoryDTO categoryDTO) {
+        if(categoryDTO.getId()==null){
+            throw new BusinessRuleException(BusinessMessages.parameterCanNotEmpty);
+        }
         categoryService.updateCategory(categoryDTO);
         return categoryDTO;
     }
 
     @GetMapping("/list/{id}")
     public List<ProductDTO> getProductsCategoryById(@PathVariable Long id) {
+        if(id== null){
+            throw new BusinessRuleException(BusinessMessages.idCanNotEmpty);
+        }
         return categoryService.getProductsCategoryById(id);
     }
 
     @GetMapping("/{id}")
     public CategoryDTO getCategoryById(@PathVariable Long id) {
+        if(id== null){
+            throw new BusinessRuleException(BusinessMessages.idCanNotEmpty);
+        }
         return categoryService.getCategoryById(id);
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteCategory(@PathVariable long id) {
+    public Long deleteCategory(@PathVariable Long id) {
+        if(id== null){
+            throw new BusinessRuleException(BusinessMessages.idCanNotEmpty);
+        }
         categoryService.deleteCategory(id);
         return null;
     }
