+++ src/main/java/com/ba/restaurant/service/TableCategoryService.java
@@ -1,14 +1,17 @@
 package com.ba.restaurant.service;
 
+import com.ba.restaurant.config.LocaleConfig;
 import com.ba.restaurant.dto.TableCategoryDTO;
 import com.ba.restaurant.entity.TableCategory;
+import com.ba.restaurant.exception.BusinessMessages;
+import com.ba.restaurant.exception.SystemException;
 import com.ba.restaurant.mapper.TableCategoryMapper;
 import com.ba.restaurant.repository.TableCategoryRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Locale;
 
 @Service
 public class TableCategoryService {
@@ -16,34 +19,47 @@ public class TableCategoryService {
     @Autowired
     TableCategoryRepository tableCategoryRepository;
 
-    public TableCategoryDTO addTableCategory(TableCategoryDTO tableCategoryDTO) {
-        TableCategory tableCategory = TableCategoryMapper.INSTANCE.toEntity(tableCategoryDTO);
+    @Autowired
+    TableCategoryMapper tableCategoryMapper;
+
+    public TableCategoryDTO addTableCategory(TableCategoryDTO tableCategoryDTO, String locale) {
+        if (tableCategoryDTO == null) {
+            throw new SystemException(LocaleConfig.messageSource().getMessage(BusinessMessages.canNotBeAdded, null, new Locale(locale)));
+        }
+        TableCategory tableCategory = tableCategoryMapper.toEntity(tableCategoryDTO);
         tableCategoryRepository.save(tableCategory);
         return tableCategoryDTO;
     }
 
-    public TableCategoryDTO updateTableCategory(TableCategoryDTO tableCategoryDTO) {
-        TableCategory tableCategory = TableCategoryMapper.INSTANCE.toEntity(tableCategoryDTO);
+    public TableCategoryDTO updateTableCategory(TableCategoryDTO tableCategoryDTO, String locale) {
+        if (tableCategoryDTO == null) {
+            throw new SystemException(LocaleConfig.messageSource().getMessage(BusinessMessages.canNotBeUpdated, null, new Locale(locale)));
+        }
+        TableCategory tableCategory = tableCategoryMapper.toEntity(tableCategoryDTO);
         tableCategoryRepository.saveAndFlush(tableCategory);
         return tableCategoryDTO;
     }
 
-    public TableCategoryDTO getTableCategorytById(Long id) {
+    public TableCategoryDTO getTableCategorytById(Long id, String locale) {
+        if (id == null) {
+            throw new SystemException(LocaleConfig.messageSource().getMessage(BusinessMessages.idCanNotfound, null, new Locale(locale)));
+        }
         TableCategory tableCategory = tableCategoryRepository.findById(id).get();
-        TableCategoryDTO tableCategoryDTO = TableCategoryMapper.INSTANCE.toDTO(tableCategory);
-        return tableCategoryDTO;
+        return tableCategoryMapper.toDTO(tableCategory);
     }
 
     public List<TableCategoryDTO> listAllTableCategory() {
-        List<TableCategoryDTO> tableCategoryDTOS = new ArrayList<>();
         List<TableCategory> tableCategories = tableCategoryRepository.findAll();
-        tableCategories.forEach(tableCategory -> tableCategoryDTOS.add(TableCategoryMapper.INSTANCE.toDTO(tableCategory)));
-        return tableCategoryDTOS;
+        return tableCategoryMapper.toDTOS(tableCategories);
     }
 
-    public Long deleteByTableCategory(Long id) {
+    public Long deleteByTableCategory(Long id, String locale) {
+        if (id == null) {
+            throw new SystemException(LocaleConfig.messageSource().getMessage(BusinessMessages.idCanNotfound, null, new Locale(locale)));
+        }
         tableCategoryRepository.deleteById(id);
-        return id;
+        return null;
     }
 
 }
+
