+++ src/main/java/com/ba/restaurant/service/WaiterService.java
@@ -2,13 +2,14 @@ package com.ba.restaurant.service;
 
 import com.ba.restaurant.dto.WaiterDTO;
 import com.ba.restaurant.entity.Waiter;
+import com.ba.restaurant.exception.BusinessMessages;
+import com.ba.restaurant.exception.SystemException;
 import com.ba.restaurant.mapper.WaiterMapper;
 import com.ba.restaurant.repository.MediaRepository;
 import com.ba.restaurant.repository.WaiterRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,33 +22,42 @@ public class WaiterService {
     @Autowired
     MediaRepository mediaRepository;
 
+    @Autowired
+    WaiterMapper waiterMapper;
+
     public WaiterDTO addWaiter(WaiterDTO waiterDTO) {
-        Waiter waiter = WaiterMapper.INSTANCE.toEntity(waiterDTO);
-        waiterRepository.save(waiter);
+        if (waiterDTO == null) {
+            throw new SystemException(BusinessMessages.canNotBeAdded);
+        }
+        waiterRepository.save(waiterMapper.toEntity(waiterDTO));
         return waiterDTO;
     }
 
     public WaiterDTO updateWaiter(WaiterDTO waiterDTO) {
-        Waiter waiter = WaiterMapper.INSTANCE.toEntity(waiterDTO);
-        waiterRepository.saveAndFlush(waiter);
+        if (waiterDTO == null) {
+            throw new SystemException(BusinessMessages.canNotBeUpdated);
+        }
+        waiterRepository.saveAndFlush(waiterMapper.toEntity(waiterDTO));
         return waiterDTO;
     }
 
     public List<WaiterDTO> listAllWaiter() {
-        List<WaiterDTO> waiterDTOS = new ArrayList<>();
         List<Waiter> waiters = waiterRepository.findAll();
-        waiters.forEach(waiter -> waiterDTOS.add(WaiterMapper.INSTANCE.toDTO(waiter)));
-        return waiterDTOS;
+        return waiterMapper.toDTOs(waiters);
     }
 
     public WaiterDTO getWaiterById(Long id) {
+        if (id == null) {
+            throw new SystemException(BusinessMessages.idCanNotfound);
+        }
         Optional<Waiter> waiter = waiterRepository.findById(id);
-        WaiterDTO waiterDTO = new WaiterDTO();
-        waiterDTO = WaiterMapper.INSTANCE.toDTO(waiter.get());
-        return waiterDTO;
+        return waiterMapper.toDTO(waiter.get());
     }
 
     public String deleteWaiterById(Long id) {
+        if (id == null) {
+            throw new SystemException(BusinessMessages.idCanNotfound);
+        }
         waiterRepository.deleteById(id);
         return null;
     }
